refactor: Split setup.ts into separate route files
- Created src/routes/setup/ directory with individual route handlers:
  - google-tokens.ts: DELETE /google-tokens
  - google-config.ts: POST /google-config
  - google-auth.ts: GET /google-auth
  - google-callback.ts: GET /google-callback
  - sheets.ts: GET /sheets
  - initialize-sheet-stream.ts: POST /initialize-sheet-stream
  - initialize-sheet.ts: POST /initialize-sheet
  - complete.ts: POST /complete

- Simplified main setup.ts from 680+ lines to 50 lines
- All route handlers maintain exact same functionality
- Improved maintainability and code organization
