+++ src/routes/setup/google-tokens.ts
@@ -1,0 +1,28 @@
+/**
+ * DELETE /api/setup/google-tokens
+ *
+ * Clear stored Google OAuth tokens (for re-authentication)
+ */
+
+import type { Context } from 'hono';
+import type { Env } from '../../types/env';
+import { ConfigRepository } from '../../db/config.repository';
+
+export async function deleteGoogleTokens(c: Context<{ Bindings: Env }>) {
+  try {
+    const configRepo = new ConfigRepository(c.env);
+
+    // Delete all Google-related tokens from config
+    await configRepo.set('google_access_token', '', 'Cleared');
+    await configRepo.set('google_refresh_token', '', 'Cleared');
+    await configRepo.set('google_token_expires_at', '', 'Cleared');
+
+    return c.json({ success: true, message: 'Tokens cleared' });
+  } catch (error) {
+    console.error('Error clearing tokens:', error);
+    return c.json(
+      { error: error instanceof Error ? error.message : 'Failed to clear tokens' },
+      500
+    );
+  }
+}
